rename scene_parts to elements
