+++ crates/re_space_view_spatial/src/scene/elements/mod.rs
